_site/requirements.txt
Django==1.3
-e hg+ssh://hg@bitbucket.org/jespern/django-piston@c4b2d21db51a531b069dd7f0d7fc10996c9c70c8#egg=django_piston-tip
django-social-auth==0.4.2
httplib2==0.7.1
-e git+git@github.com:felixleong/python-oauth2.git@1fcc1a6d49a7be2a9394125acf5b1c2227fcde79#egg=oauth2-1.5.170-py2.7-dev
python-openid==2.2.5
wsgiref==0.1.2
